third_party/WebKit/LayoutTests/inspector-protocol/page/frameAttachedStacktrace-expected.txt

Tests stackTrace reported on frame attach.
Frame Attached
Stack is empty

Frame Attached
Call Frames :
[
  [0] : {
    columnNumber : 18
    functionName : createFrame
    lineNumber : 13
    scriptId : <scriptId>
    lineNumber : 13
    url : <only showing file name>/frame-attached-stacktrace-page.html
  }
  [1] : {
    columnNumber : 0
    functionName : 
    lineNumber : 0
    scriptId : <scriptId>
    lineNumber : 0
    url : 
  }
]

Frame Attached
Call Frames :
[
  [0] : {
    columnNumber : 14
    functionName : 
    lineNumber : 3
    scriptId : <scriptId>
    lineNumber : 3
    url : data:text/html,<html><body><script>%0Avar innerFrame = document.createElement('iframe');%0AinnerFrame.src = 'about:blank';%0Adocument.body.appendChild(innerFrame);%0A</script></body></html>
  }
]
